extract runPreviewSwarm to it's own fn so we can call it in e2e tests
